Clean up BRAK-IdP OpenID spike
