+++ .env.example
@@ -1,8 +1,6 @@
-BASE_URL=http://localhost:3000
-
+BASE_URL=
 # get BRAK IdP values from 1Password by searching for "BRAK IdP (SPT)"
-BRAK_IDP_OIDC_AUDIENCE=
 BRAK_IDP_OIDC_CLIENT_ID=
 BRAK_IDP_OIDC_CLIENT_SECRET=
 BRAK_IDP_OIDC_ISSUER=
-BRAK_IDP_OIDC_REDIRECT_BASE=
+BRAK_IDP_OIDC_REDIRECT_URI=
