+++ app/routes/_index.tsx
@@ -1,10 +1,11 @@
 import type {
   ActionFunctionArgs,
+  LoaderFunction,
   LoaderFunctionArgs,
   MetaFunction,
 } from "@remix-run/node";
-import { data, Form, redirect } from "@remix-run/react";
-import { requireUserSession } from "~/services/session.server";
+import { data, Link } from "@remix-run/react";
+import { getSession } from "~/services/session.server";
 
 export const meta: MetaFunction = () => {
   return [
@@ -13,31 +14,19 @@ export const meta: MetaFunction = () => {
   ];
 };
 
-// @TODO:
-// - test user auth
-// - add error handling
-export async function action({ request }: ActionFunctionArgs) {
-  // let session = await sessionStorage.getSession(request.headers.get("cookie"));
+// TODO:
+// - test user auth with BRAK IdP test env "schulung"
+// - auth flow: add error handling
+// - on page load: if user is logged in already, redirect to /dashboard
+//   - authorizeUser() via auth.callback.tsx, when redirect_url has been updated @ BRAK
 
-  await requireUserSession(request)
-    .then((data) => {
-      console.log("requireUserSession then", data);
-    })
-    .catch((error) => {
-      console.log("requireUserSession catch", error);
-      throw redirect("/error");
-    });
-
-  throw redirect("/dashboard", {
-    // headers: { "Set-Cookie": await sessionStorage.commitSession(session) },
-  });
-}
-
-// @TODO: if user is logged in already, redirect to /dashboard
 export async function loader({ request }: LoaderFunctionArgs) {
-  // let session = await sessionStorage.getSession(request.headers.get("cookie"));
-  // let user;
-  // if (user) throw redirect("/dashboard");
+  // debugging:
+  const session = await getSession(request.headers.get("Cookie"));
+  const codeVerifier = session.get("code_verifier");
+  const return_to = session.get("return_to");
+  console.log("codeVerifier is", codeVerifier);
+  console.log("return_to is", return_to);
 
   return data(null);
 }
@@ -52,12 +41,13 @@ export default function Index() {
         Willkommen auf der Pilotplattform für den digitalen Austausch zwischen
         Gerichten und Verfahrensbeteiligten in Zivilprozessen vor Amtsgerichten
       </h2>
-      <Form method="post">
-        <div className={"m-40 text-center"}>
-          <p className={"pb-20"}>Bitte wählen Sie Ihre Loginmethode:</p>
-          <button className={"ds-button"}>beA-Portal</button>
-        </div>
-      </Form>
+
+      <div className={"m-40 text-center"}>
+        <p className={"pb-20"}>Bitte wählen Sie Ihre Loginmethode:</p>
+        <Link to={"/login/bea"} className={"ds-button"}>
+          beA-Portal
+        </Link>
+      </div>
     </main>
   );
 }
