+++ app/routes/auth.callback.tsx
@@ -1,19 +1,28 @@
-import { data, type LoaderFunction } from "@remix-run/node";
-import { authorizeUser } from "~/services/session.server";
+import { data, redirect, type LoaderFunction } from "@remix-run/node";
+import { authorizeUser } from "~/services/brakAuth.server";
+
+// needs to be done on root URL at the moment (redirect_url update needed by BRAK)
 
 export const loader: LoaderFunction = async ({ request }) => {
-  const hi = await authorizeUser(request);
-  console.log("callback route has been called");
+  await authorizeUser(request)
+    .then((data) => {
+      console.log("authorizeUser then", data);
+      throw redirect("/dashboard");
+    })
+    .catch((error) => {
+      console.log("authorizeUser catch", error);
+      throw redirect("/error");
+    });
 
   return data(null);
 };
 
 export default function AuthCallback() {
-  // @TODO: show something helpful for auth issues
-
   return (
     <main className={"m-40 flex flex-col items-center"}>
-      <h1 className={"ds-heading-01-bold mb-40 break-all"}>Callback error!?</h1>
+      <h1 className={"ds-heading-01-bold mb-40 break-all"}>
+        Authentication callback error
+      </h1>
     </main>
   );
 }
