+++ app/routes/login.bea.tsx
@@ -1,0 +1,9 @@
+import type { LoaderFunction, MetaFunction } from "@remix-run/node";
+import { requireUserSession } from "~/services/brakAuth.server";
+
+export const loader: LoaderFunction = async ({ request }) => {
+  console.log("BeaLogin loader...");
+
+  await requireUserSession(request);
+  return null;
+};
